gpu: scaffold render funcs
The idea is to separate rendering into smaller phases: background, then window, then sprites.

The next step for rendering the background layer is to refactor the vram. Whenever we write
to tile map data, change the color palette, or redefine the tiles we should recalculate the
per pixel matrix representing the screen.

This background data should look like a 256x256 pixel array. Then we can scope it to the gb
screen size (160x144) using the scroll offsets to produce the final image.